Publish plugin via CI
If the branch is protected (i.e. `main`) and has a git tag starting with
`v` (i.e. `v1.2.3`), the plugin will be published.

First get the publish key and secret from the [Gradle
Plugins](https://plugins.gradle.org/) page.

Then set the CI variables for the plugins on the [repository settings
page](https://gitlab.com/SalvatoreT/uniffi-kotlin-multiplatform-bindings/-/settings/ci_cd).

The key should be `GRADLE_PUBLISH_KEY` and the secret should be
`GRADLE_PUBLISH_SECRET`.

After this commit is on `main`, tag the HEAD with a version and push the
tag up.

Closes #60