Don't build the dependency image in PR checks (#30)
GitHub doesn't allow workflows to have the write packages permissions
when they are triggered by the pull_request event. An error happens when
the PR check workflow tries to push a new dependency image.

This PR makes `dependency-image.yml` don't push the new image in PRs.
When a new image is needed, it will return the `:pr-latest` tag instead.
When `:pr-latest` is returned, `pr-build-test.yml` will build but not
push the image. This prevents the permission error.

When the PR is merged, `dependency-image.yml` will push the new image.