Update the test status badge (#69)
## Changes

The "check runs" is registering as "failed", but the build and test
actions have all been passing. To have the badge reflect the current
test status, let's use the result of the "Build and run tests" action.

As a bonus, link to the "Build and run tests" page.

### Before (no link)

![Gitlab Build
Status](https://img.shields.io/github/check-runs/gobley/gobley/main
)

### After (with link)

[![GitHub Actions Workflow Status](

https://img.shields.io/github/actions/workflow/status/gobley/gobley/pr-build-test.yml?branch=main&label=tests)](

https://github.com/gobley/gobley/actions/workflows/pr-build-test.yml?query=branch%3Amain)