Add DSL for configuring UniFFI in Gradle (#67)
## Changes

- Made `generateFromLibrary` default. When the user doesn't explicitly
set this in the `uniffi {}` block, the UniFFI plugin will generate
bindings from the Rust library.
- Added properties for UniFFI configuration to `BindingsGeneration`,
which is then propagated to `MergeUniffiConfigTask`.

## Testing

- Replaced `uniffi.toml` in `:tests:uniffi:coverall` and
`:examples:custom-types` with Gradle DSL.
- Removed `:tests:uniffi:no-uniffi-block`.

## Additional context

This is another PR that will improve the experience of newcomers. This
allows using UniFFI without making additional files like `uniffi.toml`
or `<namespace>.udl`. The UniFFI configuration file can be modified via
type-safe DSL on the Gradle side.

Scenarios like `:tests:uniffi:no-uniffi-block`, which applies the UniFFI
plugin without actually using UniFFI, will be unsupported. Gradle
plugins like `application`, AGP, or the Cargo plugin do their logic
right after when they're applied, so changing the behavior of the UniFFI
plugin like this wouldn't be that destructive.

Determining the default value of `cdylibName` in the UniFFI plugin is
redundant because:
- It should be set to another value when using the external type
feature.
- When the value is not present on the Gradle side, UniFFI determines it
on the Rust side.