Ease version requirements of Kotlin dependencies by UniFFI (#89)
## Changes

- Added `UniFfiExtension.addDependencies: Property<Boolean>`. When this
is `false`, the UniFFI plugin doesn't apply the required dependencies.
This defaults to `true` (`true` means that the plugin keeps the previous
behavior).
- Made version requirements of Kotlin dependencies set by the UniFFI
plugin use `prefer` requirements, which can be overridden by other
requirements set by the user.
- For JNA on Android, we need two different artifacts (the `.jar` one
and the `.aar` one), so the UniFFI plugin tries retrieving the JNA
version and uses the default version (`DependencyVersions.JNA`) only if
the user doesn't specify the JNA version manually. The version selected
by this process is used both in the main sourceSet and the local unit
tests.

## Testing

Manually tested with the tutorials added by #62.

## Issues Fixed

Fixes: #15