Propagate desktop JVM dependencies only when invoked by Compose previews (#94)
## Changes

- Added `GradleUtils.getComposePreviewVariant()`, which returns a
`Variant` when the build was invoked by Compose previews.
- It returns `Variant.Debug` during IDE sync, which effectively prevents
support for release mode Compose previews. An issue for this is opened
at #99.
- It *returns* `Variant.Release` when invoked from release mode Compose
previews, but since during IDE sync the JAR file for the debug build is
used, the release preview tries to find the debug mode Rust dynamic
library.
  - It returns `null` during normal builds.
- The Rust dynamic library JAR file and the desktop JNA are used inside
`runtimeOnly` only if `GradleUtils.getComposePreviewVariant()` returns
`Variant.Debug`. This makes them removed during normal builds.
- Made `DependencyUtils.resolveUniFfiDependencies` always register JNA
as a dependency, as it is used inside Compose previews or unit tests so
it is not included in the final app. This is to prevent resolving
dependencies during configuration.

## Testing

Tested with the Android tutorial added by #62. Compose test code is as
follows.

```kotlin
Column(Modifier.safeContentPadding()) {
    val url = this::class.java.classLoader
        ?.getResource("darwin-aarch64/libapp.dylib")
        ?.toString()
    if (url == null) {
        Text("NOT FOUND")
    } else {
        Text(
            AnnotatedString.Builder().apply {
                append(url)
                val debugIdx = url.indexOf("debug.jar")
                if (debugIdx != -1) {
                    addStyle(SpanStyle(color = Color.Red), debugIdx, debugIdx + 5)
                }
            }.toAnnotatedString()
        )
    }
    Text("BuildConfig.DEBUG: ${BuildConfig.DEBUG}")
    Text("JNA Version: ${Native.VERSION}")
    Text("Addition using Rust: 2 + 3 = ${uniffi.app.add(2, 3)}")
    val greeting = remember { uniffi.app.Greeter("Hello") }
    Text(greeting.greet("Rust"))
    DisposableEffect(greeting) {
        onDispose {
            greeting.close()
        }
    }
}
```

Tested with a multi-module project as well.

## Issues Fixed

Fixes: #92